Pharmacy_Shop/pyproject.toml
[tool.poetry]
name = "pharmacy-shop"
version = "0.1.0"
description = " "
authors = ["Zekhz <Zekhz07@gmail.com>"]


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.0"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.20"
# Email format
email-validator = "^2.0.0.post2"
# For encryption and authentication
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptograpy"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyproject-toml = "^0.0.10"


[build-system]
requires = ["poetry-core >= 1.2.0"]
build-backend = "poetry.core.masonry.api"
